Rename 'BLAST+DIAMOND' to 'DIAMOND' in README
This link points to the DIAMOND software repository. The only thing that DIAMOND has in common with BLAST is that they both are designed to search for homologs in a sequence database.

You might consider adding MMSeqs2 (https://github.com/soedinglab/MMseqs2). It is much more sensitive than DIAMOND and still quite fast.